+++ packages/function/test/vm.js
@@ -1,3 +1,5 @@
+/* eslint-disable no-eval */
+
 'use strict'
 
 const getBrowserless = require('browserless')
@@ -118,11 +120,11 @@ test('run browserless code', async t => {
   const url = 'https://example.com'
   const scriptPath = path.resolve(__dirname, 'vm.js')
 
-  const code = `async (page) => page.title();`
+  const code = 'async (page) => page.title();'
 
   const template = `async ({ browserWSEndpoint, code, url, opts }) => {
     const getBrowserless = require('browserless')
-    const browserless = getBrowserless({ mode: 'connect', browserWSEndpoint })
+    const browserless = await getBrowserless({ mode: 'connect', browserWSEndpoint }).createContext()
     const browserFn = browserless.evaluate(${eval(code)})
     const result = await browserFn(url, opts)
     return result
@@ -131,16 +133,7 @@ test('run browserless code', async t => {
   const vm = createVm({
     require: {
       external: {
-        builtin: ['path', 'url'],
-        modules: [
-          'p-reflect',
-          'p-retry',
-          'browserless',
-          'metascraper',
-          'metascraper-',
-          '@metascraper',
-          'lodash'
-        ]
+        modules: ['p-reflect', 'p-retry', 'metascraper', 'metascraper-', '@metascraper', 'lodash']
       }
     }
   })